+++ crates/net/eth-wire/Cargo.toml
@@ -11,7 +11,7 @@ repository.workspace = true
 [dependencies]
 # reth
 reth-codecs = { path = "../../storage/codecs" }
-reth-primitives.workspace = true
+reth-primitives = { workspace = true, features = ["c-kzg"] }
 reth-ecies = { path = "../ecies" }
 alloy-rlp = { workspace = true, features = ["derive"] }
 reth-discv4 = { path = "../discv4" }
