+++ crates/primitives/src/lib.rs
@@ -64,6 +64,7 @@ pub use constants::{
     DEV_GENESIS_HASH, EMPTY_OMMER_ROOT_HASH, GOERLI_GENESIS_HASH, HOLESKY_GENESIS_HASH,
     KECCAK_EMPTY, MAINNET_GENESIS_HASH, SEPOLIA_GENESIS_HASH,
 };
+#[cfg(feature = "c-kzg")]
 pub use eip4844::{calculate_excess_blob_gas, kzg_to_versioned_hash};
 pub use forkid::{ForkFilter, ForkHash, ForkId, ForkTransition, ValidationError};
 pub use genesis::{Genesis, GenesisAccount};
@@ -84,12 +85,18 @@ pub use receipt::{Receipt, ReceiptWithBloom, ReceiptWithBloomRef, Receipts};
 pub use serde_helper::JsonU256;
 pub use snapshot::SnapshotSegment;
 pub use storage::StorageEntry;
+
+#[cfg(feature = "c-kzg")]
+pub use transaction::{
+    BlobTransaction, BlobTransactionSidecar, BlobTransactionValidationError,
+    FromRecoveredPooledTransaction, PooledTransactionsElement,
+    PooledTransactionsElementEcRecovered,
+};
+
 pub use transaction::{
     util::secp256k1::{public_key_to_address, recover_signer, sign_message},
-    AccessList, AccessListItem, BlobTransaction, BlobTransactionSidecar,
-    BlobTransactionValidationError, FromRecoveredPooledTransaction, FromRecoveredTransaction,
-    IntoRecoveredTransaction, InvalidTransactionError, PooledTransactionsElement,
-    PooledTransactionsElementEcRecovered, Signature, Transaction, TransactionKind, TransactionMeta,
+    AccessList, AccessListItem, FromRecoveredTransaction, IntoRecoveredTransaction,
+    InvalidTransactionError, Signature, Transaction, TransactionKind, TransactionMeta,
     TransactionSigned, TransactionSignedEcRecovered, TransactionSignedNoHash, TxEip1559, TxEip2930,
     TxEip4844, TxHashOrNumber, TxLegacy, TxType, TxValue, EIP1559_TX_TYPE_ID, EIP2930_TX_TYPE_ID,
     EIP4844_TX_TYPE_ID, LEGACY_TX_TYPE_ID,
@@ -125,6 +132,7 @@ pub type H64 = B64;
 pub use arbitrary;
 
 /// EIP-4844 + KZG helpers
+#[cfg(feature = "c-kzg")]
 pub mod kzg {
     pub use c_kzg::*;
 }
