+++ crates/primitives/src/transaction/pooled.rs
@@ -1,5 +1,8 @@
 //! Defines the types for blob transactions, legacy, and other EIP-2718 transactions included in a
 //! response to `GetPooledTransactions`.
+#![cfg(feature = "c-kzg")]
+#![cfg_attr(docsrs, doc(cfg(feature = "c-kzg")))]
+
 use crate::{
     Address, BlobTransaction, Bytes, Signature, Transaction, TransactionSigned,
     TransactionSignedEcRecovered, TxEip1559, TxEip2930, TxHash, TxLegacy, B256, EIP4844_TX_TYPE_ID,
